backend/requirements.txt
fastapi==0.95.1
pydantic==1.10.7
uvicorn[standard]==0.22.0
sqlalchemy==2.0.12
python-dotenv==1.0.0
psycopg2-binary==2.9.6
aiosqlite>=0.17.0  # For SQLite support
google-api-python-client==2.86.0
requests==2.30.0
transformers==4.30.2
torch==2.1.0
numpy>=1.24.3 
websockets==11.0.2
setuptools>=42
wheel
build
psutil>=5.9.0
httpx==0.25.0  # HTTP client with async support
Pillow==10.2.0
svgwrite>=1.4.3
python-multipart==0.0.6
pandas==2.1.0
pytz==2023.3
plotly==5.18.0
beautifulsoup4==4.12.2
google-trans-new==1.1.9
scikit-learn==1.2.2
nltk==3.8.1
kagglehub==0.2.6
gunicorn==20.1.0
tensorflow==2.15.0
cryptography==42.0.2
asyncpg
aiohttp
kaggle
Jinja2==3.1.2